refactor: extract xtdb dataframe ops
