+++ polars_hist_db/backends/xtdb_staging.py
@@ -17,9 +17,8 @@ from ..pipeline_projection import project_staged_pipeline_item_dataframe
 from ..types import PolarsType
 from .xtdb_arrow import _xtdb_json_safe_key_value
 from .xtdb_query import _xtdb_table_query_target_column, _xtdb_values_cte
+from .xtdb_dataframe import XtdbAdbcDataframeOps, XtdbDataframeOps
 from .xtdb_transport import (
-    XtdbAdbcDataframeOps,
-    XtdbDataframeOps,
     _is_xtdb_adbc_ingest_unavailable,
     _normalize_xtdb_timestamp_columns,
     _qualified_table_name,
